requirements.txt
django
pytest-django
factory-boy
Pillow
